refactor: update import statements with new module name
