+++ backend/internal/api/v1/handlers/signin.go
@@ -4,8 +4,8 @@ import (
 	"net/http"
 
 	"github.com/gin-gonic/gin"
-	"github.com/wigit-gh/webapp/internal/api/v1/middlewares"
-	"github.com/wigit-gh/webapp/internal/db"
+	"github.com/wigit-gh/webapp/backend/internal/api/v1/middlewares"
+	"github.com/wigit-gh/webapp/backend/internal/db"
 	"golang.org/x/crypto/bcrypt"
 )
