+++ backend/internal/api/v1/routes/bookings.go
@@ -2,7 +2,7 @@ package routes
 
 import (
 	"github.com/gin-gonic/gin"
-	"github.com/wigit-gh/webapp/internal/api/v1/handlers"
+	"github.com/wigit-gh/webapp/backend/internal/api/v1/handlers"
 )
 
 // AdminBookingsRoutes adds all the routes for the bookings endpoint used by the admin.
