+++ backend/internal/db/database.go
@@ -5,8 +5,8 @@ import (
 	// _ "time/tzdata" if we make use of select time zone we may have to uncomment this and move to main
 
 	"github.com/rs/zerolog/log"
-	"github.com/wigit-gh/webapp/internal/config"
-	"github.com/wigit-gh/webapp/internal/logger"
+	"github.com/wigit-gh/webapp/backend/internal/config"
+	"github.com/wigit-gh/webapp/backend/internal/logger"
 	"gorm.io/driver/mysql"
 	"gorm.io/gorm"
 )
